@@ -5,7 +5,7 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.5.4</version>
+    <version>3.3.5</version>
     <relativePath/> <!-- lookup parent from repository -->
   </parent>
   <groupId>com.nextgen</groupId>
@@ -14,7 +14,7 @@
   <name>banking-app</name>
   <description>Banking app</description>
   <properties>
-    <java.version>1.8</java.version>
+    <java.version>21</java.version>
   </properties>
   <dependencies>
     <dependency>
