@@ -1,22 +1,27 @@
 package com.nextgen.bankingapp.security;
 
-import lombok.extern.log4j.Log4j2;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.annotation.web.configurers.HeadersConfigurer;
+import org.springframework.security.web.SecurityFilterChain;
+
+import lombok.extern.log4j.Log4j2;
 
 @Configuration
 @EnableWebSecurity
 @Log4j2
-public class DefaultWebSecurityConfiguration extends WebSecurityConfigurerAdapter {
+public class DefaultWebSecurityConfiguration {
 
-  @Override
-  protected void configure(final HttpSecurity http) throws Exception {
-    http.csrf().disable().headers().frameOptions().disable().and()
-        .authorizeRequests()
-        .anyRequest()
-        .permitAll();
+  @Bean
+  protected SecurityFilterChain configure(final HttpSecurity http) throws Exception {
+    return http
+            .csrf(AbstractHttpConfigurer::disable)
+            .headers(h -> h.frameOptions(HeadersConfigurer.FrameOptionsConfig::disable))
+            .authorizeHttpRequests(auth -> auth.anyRequest().permitAll())
+            .build();
   }
 
 }
