@@ -1,15 +1,17 @@
 package com.nextgen.bankingapp.services.database;
 
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.Entity;
-import javax.persistence.Id;
 import java.io.Serializable;
 
 @Entity
 @Getter
 @Setter
+@Table(name = "users")
 public class User implements Serializable {
   @Id
   private String username;
